Banker Roulette program 1
This is the most simple program demonstrating the random module, where the user would have to input certain names of individual and the output would be printed with a random names from the list who would be required to pay the food bills. This program is based on the method where a group of friends visit hotel for food, and at the time of bill payment, they put their card on the table for the waiter to choose. Based on the random selection, one of the card got selected and the individual would have to pay the bill